+++ alpha/lib/enums/RuleActionType.php
@@ -11,4 +11,5 @@ interface RuleActionType extends BaseEnum
 	const ADD_TO_STORAGE = 4;
 	const LIMIT_DELIVERY_PROFILES = 5;
 	const SERVE_FROM_REMOTE_SERVER = 6;
+	const REQUEST_HOST_REGEX = 7;
 }