+++ deployment/updates/sql/2015_08_31_alter_category_name_column.sql
@@ -1,0 +1,1 @@
+ALTER TABLE category modify name text NOT NULL;