+++ plugins/content_distribution/providers/tvinci/lib/admin/TvinciProfileConfiguration.php
@@ -5,11 +5,6 @@
  */
 class Form_TvinciProfileConfiguration extends Form_ConfigurableProfileConfiguration
 {
-	/**
-	 * This element id is used to separate the default profile elements and tvinci profile elements, so later we could
-	 * insert the dynamic elements at the correct position
-	 */
-	const FORM_PLACEHOLDER_ELEMENT_ID = 'tvinci_placeholder';
 
 	public function init()
 	{
@@ -24,6 +19,17 @@ class Form_TvinciProfileConfiguration extends Form_ConfigurableProfileConfigurat
 	public function getObject($objectType, array $properties, $add_underscore = true, $include_empty_fields = false)
 	{
 		$object = parent::getObject($objectType, $properties, $add_underscore, $include_empty_fields);
+
+		if($object instanceof Kaltura_Client_TvinciDistribution_Type_TvinciDistributionProfile) {
+			$upload = new Zend_File_Transfer_Adapter_Http();
+			$files = $upload->getFileInfo();
+
+			if (isset($files['xsltFile'])) {
+				$file = $files['xsltFile'];
+				$content = file_get_contents($file['tmp_name']);
+				$object->xsltFile = $content;
+			}
+		}
 		return $object;
 	}
 
@@ -35,29 +41,45 @@ class Form_TvinciProfileConfiguration extends Form_ConfigurableProfileConfigurat
 		$element->setLabel('Tvinci Specific Configuration');
 		$element->setDecorators(array('ViewHelper', array('Label', array('placement' => 'append')), array('HtmlTag',  array('tag' => 'b'))));
 		$this->addElements(array($element));
-		$this->addElement('hidden', self::FORM_PLACEHOLDER_ELEMENT_ID);
-		$this->_sort();
-		$order = $this->_order[self::FORM_PLACEHOLDER_ELEMENT_ID];
-		$this->resetOrderOfLastElements();
-		$this->layoutForm($order++);
+		$this->tvinciElements();
 	}
 
-
-
-	public function resetOrderOfLastElements()
+	public function populateFromObject($object, $add_underscore = true)
 	{
-		$found = false;
-		foreach ($this->_order as $key => &$order)
-		{
-			if ($found)
-				$order = null;
+		$this->tvinciElements();
 
-			if ($key == self::FORM_PLACEHOLDER_ELEMENT_ID)
-				$found = true;
+		parent::populateFromObject($object, $add_underscore);
+
+		if ($object->xsltFile) {
+			$this->getElement('xsltFileText')->setValue(json_encode($object->xsltFile));
 		}
+
 	}
 
-	protected function layoutForm($order)
+	private function addTagItems($tagName){
+		$this->addElement('text', "{$tagName}_file_name", array(
+			'label'			=> "{$tagName} file name:",
+			'filters'		=> array('StringTrim'),
+		));
+
+		$this->addElement('text', "{$tagName}_ppv_module", array(
+			'label'			=> "{$tagName} PPv Module:",
+			'filters'		=> array('StringTrim'),
+		));
+
+		$this->addDisplayGroup(
+			array("{$tagName}_file_name","{$tagName}_ppv_module" ),
+			"{$tagName}",
+			array(
+				'legend' => "{$tagName} Configuration",
+				'decorators' => array('FormElements', 'Fieldset'),
+			)
+		);
+	}
+
+
+
+	protected function tvinciElements()
 	{
 		// Ingest Configuration
 		$this->addElement('text', 'ingest_url', array(
@@ -76,38 +98,54 @@ class Form_TvinciProfileConfiguration extends Form_ConfigurableProfileConfigurat
 		));
 
 		$this->addDisplayGroup(
-				array('ingest_url','username','password'),
+				array('ingest_url','username','password' ),
 				'ingest',
 				array(
 					'legend' => 'Ingest URL Configuration',
 					'decorators' => array('FormElements', 'Fieldset'),
-					'order' => $order++,
 				)
 		);
 
-		$this->addElement('select', 'schema_id', array(
-			'label'			=> 'Schema ID:',
-			'multioptions' => array(
-				'' => '2',
-				'1' => '1',
-				'2' => '2',
-			)
+		// tag specific configuration
+		$this->addTagItems("ism");
+		$this->addTagItems("ipadnew");
+		$this->addTagItems("iphonenew");
+		$this->addTagItems("mbr");
+		$this->addTagItems("dash");
+
+		// xslt configuration
+		$this->addElement('file', 'xsltFile', array(
+			'label' => 'XSLT:'
 
 		));
 
-		$this->addElement('text', 'language', array(
-			'label'			=> 'Language :',
-			'filters'		=> array('StringTrim'),
+		$element = new Zend_Form_Element_File('xsltFile');
+		$element->setLabel('XSLT:');
+		// limit only 1 file
+		$element->addValidator('Count', true, 1);
+		// limit to 100K (the default one is 31K)
+		$element->addValidator('Size', true, 102400);
+		// only XML related file extensions
+		$element->addValidator('Extension', true, 'xml,xslt,xsl');
+		$element->addValidator(new XSLTFileValidator());
+		$this->addElement($element, 'xsltFile');
+
+
+		$this->addElement('textarea', 'xsltFileText', array(
+			'label' => 'XSLT Data:',
+			'rows' => '2',
+			'cols' => '50',
+			'readonly' => '1'
 		));
 
 		$this->addDisplayGroup(
-			array('schema_id', 'language'),
+			array('xsltFile' , 'xsltFileText'),
 			'additional',
 			array(
 				'legend' => 'Additional Configuration',
 				'decorators' => array('FormElements', 'Fieldset'),
-				'order' => $order++,
 			)
 		);
+
 	}
 }